requirements.txt
discord.py
python-dotenv
flask
pydantic
pydantic-settings